chore: updated server structure
