+++ server/src/main.rs
@@ -1,6 +1,8 @@
 use axum::{http::Method, Extension};
 use clap::Parser;
+use std::net::SocketAddr;
 use std::sync::Arc;
+use tokio::signal;
 use tower::ServiceBuilder;
 use tower_http::{
     cors::{Any, CorsLayer},
@@ -10,16 +12,15 @@ use tower_http::{
 use crate::config::*;
 use crate::db::cache::*;
 use crate::db::conn::*;
-use crate::http::serve;
-use crate::http::Build;
 use crate::prelude::*;
+use crate::router::Build;
 
-mod chain;
 mod config;
 mod db;
 mod error;
-mod http;
+mod mina;
 mod prelude;
+mod router;
 mod types;
 
 extern crate tracing;
@@ -28,8 +29,8 @@ pub(crate) const MINA_GOVERNANCE_SERVER: &str = "mina_governance_server";
 
 #[tokio::main]
 async fn main() -> Result<()> {
-    dotenvy::dotenv().unwrap_or_else(|_| panic!("Error: .env file not found."));
-    config::init_tracing().unwrap_or_else(|_| panic!("Error: failed to initialize tracing."));
+    dotenvy::dotenv().unwrap_or_else(|_| panic!("Error: .env file not found"));
+    config::init_tracing().unwrap_or_else(|_| panic!("Error: failed to initialize tracing"));
 
     let config = Config::parse();
     let cache = CacheManager::build();
@@ -50,7 +51,42 @@ async fn main() -> Result<()> {
             })),
     );
 
-    tracing::info!(target: MINA_GOVERNANCE_SERVER, "Axum runtime starting");
+    tracing::info!(target: MINA_GOVERNANCE_SERVER, "Axum runtime starting...");
     serve(router, config.port).await;
     Ok(())
+}
+
+async fn serve(router: axum::Router, port: u16) {
+    let addr = SocketAddr::from(([127, 0, 0, 1], port));
+    axum::Server::bind(&addr)
+        .serve(router.into_make_service())
+        .with_graceful_shutdown(shutdown())
+        .await
+        .expect("Error: failed to start axum runtime");
+}
+
+async fn shutdown() {
+    let windows = async {
+        signal::ctrl_c()
+            .await
+            .unwrap_or_else(|_| panic!("Error: failed to install windows shutdown handler"));
+    };
+
+    #[cfg(unix)]
+    let unix = async {
+        signal::unix::signal(signal::unix::SignalKind::terminate())
+            .unwrap_or_else(|_| panic!("Error: failed to install unix shutdown handler"))
+            .recv()
+            .await;
+    };
+
+    #[cfg(not(unix))]
+    let terminate = std::future::pending::<()>();
+
+    tokio::select! {
+        _ = windows => {},
+        _ = unix => {},
+    }
+
+    println!("Signal received - starting graceful shutdown...");
 }
